Change GetPageSize() return type from off_t to int.
Both Windows and Unix native system call are int.

+++ include/mcld/Support/SystemUtils.h
@@ -29,7 +29,7 @@ char *strerror(int pErrnum);
 
 std::string getDefaultTargetTriple();
 
-off_t GetPageSize();
+int GetPageSize();
 
 } // namespace of sys
 } // namespace of mcld
